Move clock.bot.inline.query.result package to clock.result, as it should be framework-independent
